+++ mailet/test/src/main/java/org/apache/mailet/base/test/FakeMail.java
@@ -28,7 +28,6 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.Date;
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -435,21 +434,8 @@ public class FakeMail implements Mail {
     }
 
     @Override
-    public Serializable getAttribute(String name) {
-        return toSerializable(attributes.get(AttributeName.of(name)));
-    }
-
-    @Override
     public Optional<Attribute> getAttribute(AttributeName name) {
         return Optional.ofNullable(attributes.get(name));
-    }
-
-    @Override
-    public Iterator<String> getAttributeNames() {
-        return attributes.keySet()
-            .stream()
-            .map(AttributeName::asString)
-            .iterator();
     }
 
     @Override
@@ -463,11 +449,6 @@ public class FakeMail implements Mail {
     }
 
     @Override
-    public Serializable removeAttribute(String name) {
-        return toSerializable(attributes.remove(AttributeName.of(name)));
-    }
-
-    @Override
     public Optional<Attribute> removeAttribute(AttributeName attributeName) {
         return Optional.ofNullable(attributes.remove(attributeName));
     }
@@ -475,14 +456,6 @@ public class FakeMail implements Mail {
     @Override
     public void removeAllAttributes() {
         attributes.clear();
-    }
-
-    @Override
-    public Serializable setAttribute(String name, Serializable object) {
-        Attribute attribute = Attribute.convertToAttribute(name, object);
-        Attribute previous = attributes.put(attribute.getName(), attribute);
-
-        return toSerializable(previous);
     }
 
     private Serializable toSerializable(Attribute previous) {