+++ server/container/core/src/main/java/org/apache/james/server/core/MailImpl.java
@@ -28,7 +28,6 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.Date;
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -726,33 +725,14 @@ public class MailImpl implements Disposable, Mail {
     }
 
     @Override
-    public Serializable getAttribute(String key) {
-        return toSerializable(attributes.get(AttributeName.of(key)));
-    }
-
-    @Override
     public Optional<Attribute> getAttribute(AttributeName name) {
         return Optional.ofNullable(attributes.get(name));
-    }
-
-    @Override
-    public Serializable setAttribute(String key, Serializable object) {
-        Preconditions.checkNotNull(key, "Key of an attribute should not be null");
-        Attribute attribute = Attribute.convertToAttribute(key, object);
-        Attribute previous = attributes.put(attribute.getName(), attribute);
-
-        return toSerializable(previous);
     }
 
     @Override
     public Optional<Attribute> setAttribute(Attribute attribute) {
         Preconditions.checkNotNull(attribute.getName().asString(), "AttributeName should not be null");
         return Optional.ofNullable(this.attributes.put(attribute.getName(), attribute));
-    }
-
-    @Override
-    public Serializable removeAttribute(String key) {
-        return toSerializable(attributes.remove(AttributeName.of(key)));
     }
 
     @Override
@@ -764,14 +744,6 @@ public class MailImpl implements Disposable, Mail {
     @Override
     public void removeAllAttributes() {
         attributes.clear();
-    }
-
-    @Override
-    public Iterator<String> getAttributeNames() {
-        return attributes.keySet()
-            .stream()
-            .map(AttributeName::asString)
-            .iterator();
     }
 
     @Override